review fixes

+++ client/asset/dcr/config.go
@@ -33,12 +33,12 @@ var (
 // Config holds the parameters needed to initialize an RPC connection to a dcr
 // wallet. Default values are used for RPCListen and/or RPCCert if not set.
 type Config struct {
-	RPCUser         string `ini:"username"`
-	RPCPass         string `ini:"password"`
-	RPCListen       string `ini:"rpclisten"`
-	RPCCert         string `ini:"rpccert"`
-	UseSplitTx      bool   `ini:"txsplit"`
-	FallbackFeeRate uint64 `ini:"fallbackfee"`
+	RPCUser         string  `ini:"username"`
+	RPCPass         string  `ini:"password"`
+	RPCListen       string  `ini:"rpclisten"`
+	RPCCert         string  `ini:"rpccert"`
+	UseSplitTx      bool    `ini:"txsplit"`
+	FallbackFeeRate float64 `ini:"fallbackfee"`
 	// Context should be canceled when the application exits. This will cause
 	// some cleanup to be performed during shutdown.
 	Context context.Context `ini:"-"`
